# feat(stdlib): optimize filter to pass through tables when possible