The filter transformation has been optimized to pass through the entire
table without copying when it is possible and, in other circumstances, pass
through the constructed table immediately to the next transformation rather
than to buffer it in memory until it receives a trigger.

First, the filter transformation uses type inference to know which
attributes from the record will be accessed. If all of these attributes are
part of the group key, it knows it can evaluate the function once to
determine if it the entire table should pass or if the table should be
cleared. This means we do not have to allocate any additional memory and we
can send an empty table to the next transformation when it is filtered.

If filter cannot make this optimization, it will now construct the table
and immediately send it off. If it receives a message to retract the table,
it will pass that on to the downstream dataset. This model means that
filter has less state and it is a 1-to-1 invariant for the group key which
is much easier to reason with and optimize. There is no regrouping logic
within filter which makes this safe to do.

#1903

### Done checklist
- [x] docs/SPEC.md updated
- [x] Test cases written