@jsternberg 2019-09-23
Benchmark results:

```
goos: darwin
goarch: amd64
pkg: github.com/influxdata/flux/stdlib
BenchmarkFluxEndToEnd/filter_by_regex_compile-12         	      49	  23731636 ns/op	      3328 max_allocated/op	     12544 total_allocated/op	11075445 B/op	  212413 allocs/op
BenchmarkFluxEndToEnd/filter_by_regex_function-12        	      51	  23830058 ns/op	      3328 max_allocated/op	     14688 total_allocated/op	11324767 B/op	  220778 allocs/op
BenchmarkFluxEndToEnd/filter_by_regex-12                 	      55	  21702543 ns/op	      3328 max_allocated/op	     12544 total_allocated/op	 9957211 B/op	  193202 allocs/op
BenchmarkFluxEndToEnd/filter_by_tags-12                  	      61	  19489602 ns/op	      3328 max_allocated/op	     12544 total_allocated/op	 9282027 B/op	  181084 allocs/op
PASS
ok  	github.com/influxdata/flux/stdlib	7.284s

goos: darwin
goarch: amd64
pkg: github.com/influxdata/flux/stdlib
BenchmarkFluxEndToEnd/filter_by_regex_compile-12         	      50	  23663289 ns/op	      3328 max_allocated/op	      9728 total_allocated/op	11020138 B/op	  211242 allocs/op
BenchmarkFluxEndToEnd/filter_by_regex_function-12        	      52	  23102657 ns/op	      3328 max_allocated/op	     11872 total_allocated/op	11280123 B/op	  219648 allocs/op
BenchmarkFluxEndToEnd/filter_by_regex-12                 	      55	  21778086 ns/op	      3328 max_allocated/op	      9728 total_allocated/op	 9916548 B/op	  192232 allocs/op
BenchmarkFluxEndToEnd/filter_by_tags-12                  	      60	  19651556 ns/op	      3328 max_allocated/op	      9728 total_allocated/op	 9244328 B/op	  180198 allocs/op
PASS
ok  	github.com/influxdata/flux/stdlib	7.287s
```

The comparison which doesn't include the custom metrics (sorry, not supported yet).

```
benchmark                                             old ns/op     new ns/op     delta
BenchmarkFluxEndToEnd/filter_by_regex_compile-12      23731636      23663289      -0.29%
BenchmarkFluxEndToEnd/filter_by_regex_function-12     23830058      23102657      -3.05%
BenchmarkFluxEndToEnd/filter_by_regex-12              21702543      21778086      +0.35%
BenchmarkFluxEndToEnd/filter_by_tags-12               19489602      19651556      +0.83%

benchmark                                             old allocs     new allocs     delta
BenchmarkFluxEndToEnd/filter_by_regex_compile-12      212413         211242         -0.55%
BenchmarkFluxEndToEnd/filter_by_regex_function-12     220778         219648         -0.51%
BenchmarkFluxEndToEnd/filter_by_regex-12              193202         192232         -0.50%
BenchmarkFluxEndToEnd/filter_by_tags-12               181084         180198         -0.49%

benchmark                                             old bytes     new bytes     delta
BenchmarkFluxEndToEnd/filter_by_regex_compile-12      11075445      11020138      -0.50%
BenchmarkFluxEndToEnd/filter_by_regex_function-12     11324767      11280123      -0.39%
BenchmarkFluxEndToEnd/filter_by_regex-12              9957211       9916548       -0.41%
BenchmarkFluxEndToEnd/filter_by_tags-12               9282027       9244328       -0.41%
```

I'm expecting most of the gains for this to not be present within the current benchmarks because there's not enough data, but you can see that `total_allocated` changed decently dramatically from just this small optimization.