@jsternberg 2019-09-24
Yes, you are correct. Additionally, this change makes it so that triggering is not used by filter. It passes it downstream even when it constructs a new table. The thing that would likely be next is the ability to send the table in chunks as this filter makes it still possible to end up accidentally buffer the entire thing in memory.